Move /admin/refresh_resource route to /api/v1/admin/refresh_resource; Add authenticate_request_token_or_user method for supporting both Devise auth and token auth for certain endpoints
